Keep all session tests in the same file
